@@ -1,8 +1,7 @@
 package codesquad.fineants.domain.member.domain.dto.request;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Pattern;
-
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
