@@ -2,13 +2,12 @@ package codesquad.fineants.domain;
 
 import java.time.LocalDateTime;
 
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
