@@ -1,7 +1,7 @@
 package codesquad.fineants.domain.kis.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
+import org.springframework.boot.context.properties.bind.ConstructorBinding;
 
 import lombok.Getter;
 
