@@ -2,11 +2,10 @@ package codesquad.fineants.domain.notification.domain.entity;
 
 import java.time.LocalDateTime;
 
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-
 import codesquad.fineants.domain.member.domain.entity.Member;
 import codesquad.fineants.domain.notification.domain.entity.type.NotificationType;
+import jakarta.persistence.DiscriminatorValue;
+import jakarta.persistence.Entity;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
