@@ -4,15 +4,15 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
+import org.springframework.boot.context.properties.bind.ConstructorBinding;
 
+import codesquad.fineants.domain.member.service.WebClientWrapper;
 import codesquad.fineants.domain.oauth.client.OauthClient;
 import codesquad.fineants.domain.oauth.client.google.GoogleOauthClient;
 import codesquad.fineants.domain.oauth.client.kakao.KakaoOauthClient;
 import codesquad.fineants.domain.oauth.client.naver.NaverOauthClient;
 import codesquad.fineants.domain.oauth.decoder.GoogleIDTokenDecoder;
 import codesquad.fineants.domain.oauth.decoder.KakaoIDTokenDecoder;
-import codesquad.fineants.domain.member.service.WebClientWrapper;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
